Move time entry input provider into contexts
